Updated view_all_users.php to check for SESSION user_role and check if the role is an admin before executing a user delete query.

+++ admin/includes/view_all_users.php
@@ -61,10 +61,14 @@ if (isset($_GET['change_to_subscriber'])) {
 
 if (isset($_GET['delete'])) {
 
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+    if (isset($_SESSION['user_role'])) {
+        if ($_SESSION['user_role'] == 'admin') {
+            $the_user_id = $_GET['delete'];
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+            $delete_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
+    }
 }
 
 ?>
